adding in zshrc and making symlink change better

+++ makesymlinks.sh
@@ -8,7 +8,7 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="vimrc gvimrc bash_profile"    # list of files/folders to symlink in homedir
+files="vimrc gvimrc zshrc bash_profile"    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -24,8 +24,11 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ -e ~/.$file -a ! -h ~/.$file ]
+    then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    fi
 done
